Update service creation menu
