+++ pkg/cli/service/replace.go
@@ -2,7 +2,6 @@ package cliserv
 
 import (
 	"fmt"
-	"strings"
 
 	"github.com/containerum/chkit/pkg/context"
 	"github.com/containerum/chkit/pkg/export"
@@ -193,15 +192,9 @@ func Replace(ctx *context.Context) *cobra.Command {
 						},
 					},
 					{
-						Label: "Export service to file",
+						Label: "Exit",
 						Action: func() error {
-							var fname = activekit.Promt("Type filename: ")
-							fname = strings.TrimSpace(fname)
-							if fname != "" {
-								if err := (porta.Exporter{OutFile: fname}.Export(svc)); err != nil {
-									ferr.Printf("unable to export service:\n%v\n", err)
-								}
-							}
+							ctx.Exit(0)
 							return nil
 						},
 					},
