+++ pkg/model/service/servactive/service_constructor.go
@@ -5,7 +5,9 @@ import (
 	"strings"
 
 	"github.com/containerum/chkit/pkg/model/service"
+	"github.com/containerum/chkit/pkg/porta"
 	"github.com/containerum/chkit/pkg/util/activekit"
+	"github.com/containerum/chkit/pkg/util/ferr"
 	"github.com/containerum/chkit/pkg/util/namegen"
 	"github.com/containerum/chkit/pkg/util/text"
 	"github.com/sirupsen/logrus"
@@ -66,6 +68,19 @@ func Wizard(config ConstructorConfig) (service.Service, error) {
 						}
 						border := strings.Repeat("_", text.Width(data))
 						fmt.Printf("%s\n%s\n%s\n", border, data, border)
+						return nil
+					},
+				},
+				{
+					Label: "Export service to file",
+					Action: func() error {
+						var fname = activekit.Promt("Type filename: ")
+						fname = strings.TrimSpace(fname)
+						if fname != "" {
+							if err := (porta.Exporter{OutFile: fname}.Export(serv)); err != nil {
+								ferr.Printf("unable to export service:\n%v\n", err)
+							}
+						}
 						return nil
 					},
 				},
